refactor: rename lib folder
