+++ src/components/Search/Search.tsx
@@ -1,8 +1,8 @@
-import { SearchContext } from '../../lib/context';
+import { SearchContext } from '../../modules/context';
 import { Component, ReactNode } from 'react';
 import Button from '../Button';
 import SearchInput from '../SearchInput';
-import { ContextProps } from '../../lib/interfaces';
+import { ContextProps } from '../../modules/interfaces';
 
 type SearchProps = { updateContext: (newContext: ContextProps) => void };
