+++ src/modules/interfaces.ts
